pyproject.toml
[project]
name = "fastapi-crud-db-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.3",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.128.2",
    "pwdlib[argon2]>=0.3.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "setuptools>=82.0.0",
    "sqlmodel>=0.0.32",
    "uvicorn>=0.40.0",
]
